+++ examples/network_invest.jl
@@ -24,19 +24,19 @@ the natural gas power plant with CCS and the CO₂ storage node.
 function generate_example_network_investment()
     @info "Generate case data - Simple network example with investments"
 
-    # Define the different resources and their emission intensity in tCO2/MWh
-    NG = ResourceEmit("NG", 0.2)
-    Coal = ResourceCarrier("Coal", 0.35)
-    Power = ResourceCarrier("Power", 0.0)
-    CO2 = ResourceEmit("CO2", 1.0)
-    products = [NG, Coal, Power, CO2]
+    # Define the different resources and their emission intensity in t CO₂/MWh
+    ng = ResourceEmit("ng", 0.2)
+    coal = ResourceCarrier("coal", 0.35)
+    power = ResourceCarrier("power", 0.0)
+    co2 = ResourceEmit("CO₂", 1.0)
+    products = [ng, coal, power, co2]
 
     # Variables for the individual entries of the time structure
-    op_duration = 2 # Each operational period has a duration of 2
+    op_duration = 2 # Each operational period has a duration of 2 (hours)
     op_number = 4   # There are in total 4 operational periods
     operational_periods = SimpleTimes(op_number, op_duration)
 
-    # Each operational period should correspond to a duration of 2 h while a duration if 1
+    # Each operational period should correspond to a duration of 2 h while a duration of 1
     # of a strategic period should correspond to a year.
     # This implies, that a strategic period is 8760 times longer than an operational period,
     # resulting in the values below as "/year".
@@ -46,21 +46,26 @@ function generate_example_network_investment()
     T = TwoLevel(4, 1, operational_periods; op_per_strat)
     model = InvestmentModel(
         Dict(   # Emission cap for CO₂ in t/year and for NG in MWh/year
-            CO2 => StrategicProfile([170, 150, 130, 110]) * 1000,
-            NG => FixedProfile(1e6),
+            co2 => StrategicProfile([170, 150, 130, 110]) * 1000,
+            ng => FixedProfile(1e6),
         ),
         Dict(   # Emission price for CO₂ in EUR/t and for NG in EUR/MWh
 
-            CO2 => FixedProfile(0),
-            NG => FixedProfile(0),
+            co2 => FixedProfile(0),
+            ng => FixedProfile(0),
         ),
-        CO2,    # CO2 instance
+        co2,    # CO₂ instance
         0.07,   # Discount rate in absolute value
     )
 
     # Creation of the emission data for the individual nodes.
-    capture_data = CaptureEnergyEmissions(0.9)
     emission_data = EmissionsEnergy()
+    # Line above: `EmissionsEnergy` implies that the emissions data corresponds to
+    # emissions through fuel usage as calculated by the CO₂ intensity and efficiency.
+    capture_data = CaptureEnergyEmissions(0.9)
+    # Line above: `CaptureEnergyEmissions` implies that the emissions data corresponds
+    # to emissions through fuel usage as calculated by the CO₂ intensity and efficiency.
+    # 90 % of the CO₂ emissions are captured as given by the value 0.9.
 
     # Create the individual test nodes, corresponding to a system with an electricity demand/sink,
     # coal and nautral gas sources, coal and natural gas (with CCS) power plants and CO₂ storage.
@@ -71,33 +76,35 @@ function generate_example_network_investment()
             FixedProfile(100),          # Capacity in MW
             FixedProfile(30),           # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/MW/year
-            Dict(NG => 1),              # Output from the Node, in this case, NG
+            Dict(ng => 1),              # Output from the Node, in this case, ng
         ),
         RefSource(
             "coal source",              # Node id
             FixedProfile(100),          # Capacity in MW
             FixedProfile(9),            # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/MW/year
-            Dict(Coal => 1),            # Output from the Node, in this case, coal
+            Dict(coal => 1),            # Output from the Node, in this case, coal
         ),
         RefNetworkNode(
             "NG+CCS power plant",       # Node id
             FixedProfile(0),            # Capacity in MW
             FixedProfile(5.5),          # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/MW/year
-            Dict(NG => 2),              # Input to the node with input ratio
-            Dict(Power => 1, CO2 => 0), # Output from the node with output ratio
-            # Line above: CO2 is required as output for variable definition, but the
-            # value does not matter
+            Dict(ng => 2),              # Input to the node with input ratio
+            Dict(power => 1, co2 => 0), # Output from the node with output ratio
+            # Line above: `co2` is required as output for variable definition, but the
+            # value does not matter as it is not utilized in the model.
             [
                 capture_data,           # Additional data for emissions and CO₂ capture
                 SingleInvData(
-                    FixedProfile(600 * 1e3),  # Capex in EUR/MW
-                    FixedProfile(40),       # Max installed capacity [MW]
+                    FixedProfile(600 * 1e3),    # Capex in EUR/MW
+                    FixedProfile(40),           # Maximum installed capacity [MW]
                     SemiContinuousInvestment(FixedProfile(5), FixedProfile(40)),
                     # Line above: Investment mode with the following arguments:
-                    # 1. argument: min added capactity per sp [MW]
-                    # 2. argument: max added capactity per sp [MW]
+                    # 1. argument: minimum added capactity  per sp [MW]
+                    # 2. argument: maximum added capactity per sp [MW]
+                    # `SemiContinuousInvestment` implies that one either invests at least in
+                    # the minimum added capacity, or not at all.
                 ),
             ],
         ),
@@ -106,33 +113,37 @@ function generate_example_network_investment()
             FixedProfile(40),           # Capacity in MW
             FixedProfile(6),            # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/MW/year
-            Dict(Coal => 2.5),          # Input to the node with input ratio
-            Dict(Power => 1),           # Output from the node with output ratio
+            Dict(coal => 2.5),          # Input to the node with input ratio
+            Dict(power => 1),           # Output from the node with output ratio
             [emission_data],            # Additonal data for emissions
         ),
         RefStorage{AccumulatingEmissions}(
-            "CO2 storage",              # Node id
+            "CO₂ storage",              # Node id
             StorCapOpex(
                 FixedProfile(0),       # Charge capacity in t/h
                 FixedProfile(9.1),      # Storage variable OPEX for the charging in EUR/t
                 FixedProfile(0)         # Storage fixed OPEX for the charging in EUR/(t/h year)
             ),
             StorCap(FixedProfile(1e8)), # Storage capacity in t
-            CO2,                        # Stored resource
-            Dict(CO2 => 1, Power => 0.02), # Input resource with input ratio
-            # Line above: This implies that storing CO₂ requires Power
-            Dict(CO2 => 1),             # Output from the node with output ratio
-            # In practice, for CO₂ storage, this is never used.
+            co2,                        # Stored resource
+            Dict(co2 => 1, power => 0.02), # Input resource with input ratio
+            # Line above: This implies that storing CO₂ requires power
+            Dict(co2 => 1),             # Output from the node with output ratio
+            # Line above: In the case of `AccumulatingEmissions`, you must provide the
+            # stored resource as one of the keys. Its value does however not matter as the
+            # outlet flow value is fixed to 0.
             [
                 StorageInvData(
                     charge = NoStartInvData(
-                        FixedProfile(200 * 1e3),  # CAPEX [EUR/(t/h)]
-                        FixedProfile(60),       # Max installed capacity [EUR/(t/h)]
+                        FixedProfile(200 * 1e3),    # CAPEX [EUR/(t/h)]
+                        FixedProfile(60),           # Maximum installed capacity [EUR/(t/h)]
                         ContinuousInvestment(FixedProfile(0), FixedProfile(5)),
                         # Line above: Investment mode with the following arguments:
                         # 1. argument: min added capactity per sp [t/h]
                         # 2. argument: max added capactity per sp [t/h]
-                        UnlimitedLife(),        # Lifetime mode
+                        # `ContinuousInvestment` implies you can invest in a capacity between
+                        # 0 and 5 t/h in each strategic period
+                        UnlimitedLife(),        # Lifetime mode corresponding to no retirement of capacity
                     ),
                 ),
             ],
@@ -142,11 +153,14 @@ function generate_example_network_investment()
             OperationalProfile([20, 30, 40, 30]), # Demand in MW
             Dict(:surplus => FixedProfile(0), :deficit => FixedProfile(1e6)),
             # Line above: Surplus and deficit penalty for the node in EUR/MWh
-            Dict(Power => 1),           # Energy demand and corresponding ratio
+            Dict(power => 1),           # Energy demand and corresponding ratio
         ),
     ]
 
     # Connect all nodes with the availability node for the overall energy/mass balance
+    # NOTE: This hard coding based on indexing is error prone. It is in general advised to
+    #       use a mapping dictionary to avoid any problems when introducing new technology
+    #       nodes.
     links = [
         Direct("Av-NG_pp", nodes[1], nodes[4], Linear())
         Direct("Av-coal_pp", nodes[1], nodes[5], Linear())
@@ -160,6 +174,8 @@ function generate_example_network_investment()
     ]
 
     # Input data structure
+    # It is also explained on
+    # https://energymodelsx.github.io/EnergyModelsBase.jl/stable/library/public/case_element/
     case = Case(T, products, [nodes, links], [[get_nodes, get_links]])
     return case, model
 end
@@ -180,7 +196,7 @@ pretty_table(
         header = [:StrategicPeriod, :InvestCapacity],
     ),
 )
-@info "Invested capacity for the CO2 storage in the beginning of the
+@info "Invested capacity for the CO₂ storage in the beginning of the
 individual strategic periods"
 pretty_table(
     JuMP.Containers.rowtable(