+++ examples/sink_source_invest.jl
@@ -15,25 +15,25 @@ using PrettyTables
 using TimeStruct
 
 """
-    generate_example_ss_investment(lifemode = RollingLife; discount_rate = 0.05)
+    generate_example_ss_investment()
 
 Generate the data for an example consisting of an electricity source and sink.
 The electricity source has initially no capacity. Hence, investments are required.
 """
-function generate_example_ss_investment(lifemode = RollingLife; discount_rate = 0.05)
+function generate_example_ss_investment()
     @info "Generate case data - Simple sink-source example"
 
-    # Define the different resources and their emission intensity in tCO2/MWh
-    Power = ResourceCarrier("Power", 0.0)
-    CO2 = ResourceEmit("CO2", 1.0)
-    products = [Power, CO2]
+    # Define the different resources and their emission intensity in t CO₂/MWh
+    power = ResourceCarrier("power", 0.0)
+    co2 = ResourceEmit("CO₂", 1.0)
+    products = [power, co2]
 
     # Variables for the individual entries of the time structure
-    op_duration = 2 # Each operational period has a duration of 2
+    op_duration = 2 # Each operational period has a duration of 2c
     op_number = 4   # There are in total 4 operational periods
     operational_periods = SimpleTimes(op_number, op_duration)
 
-    # Each operational period should correspond to a duration of 2 h while a duration if 1
+    # Each operational period should correspond to a duration of 2 h while a duration of 1
     # of a strategic period should correspond to a year.
     # This implies, that a strategic period is 8760 times longer than an operational period,
     # resulting in the values below as "/year".
@@ -46,12 +46,11 @@ function generate_example_ss_investment(lifemode = RollingLife; discount_rate = 
 
     # Create the global data
     model = InvestmentModel(
-        Dict(CO2 => FixedProfile(10)),  # Emission cap for CO₂ in t/year
-        Dict(CO2 => FixedProfile(0)),   # Emission price for CO₂ in EUR/t
-        CO2,                            # CO₂ instance
-        discount_rate,                  # Discount rate in absolute value
+        Dict(co2 => FixedProfile(10)),  # Emission cap for CO₂ in t/year
+        Dict(co2 => FixedProfile(0)),   # Emission price for CO₂ in EUR/t
+        co2,                            # CO₂ instance
+        0.05,                           # Discount rate in absolute value
     )
-
 
     # The lifetime of the technology is 15 years, requiring reinvestment in the
     # 5th investment period
@@ -59,15 +58,17 @@ function generate_example_ss_investment(lifemode = RollingLife; discount_rate = 
 
     # Create the investment data for the source node
     investment_data_source = SingleInvData(
-        FixedProfile(300 * 1e3),  # capex [€/MW]
-        FixedProfile(50),       # max installed capacity [MW]
+        FixedProfile(300 * 1e3),    # CAPEX [€/MW]
+        FixedProfile(50),           # Maximum installed capacity [MW]
         ContinuousInvestment(FixedProfile(0), FixedProfile(30)),
         # Line above: Investment mode with the following arguments:
-        # 1. argument: min added capactity per sp [MW]
-        # 2. argument: max added capactity per sp [MW]
-        lifemode(lifetime),     # Lifetime mode
+        # 1. argument: minimum added capactity per sp [MW]
+        # 2. argument: maximum added capactity per sp [MW]
+        # `ContinuousInvestment` implies you can invest in a capacity between 0 and 30 MW in
+        # each strategic period
+        RollingLife(lifetime),     # Lifetime mode
+        # Line above: As default we are using `RollingLife` with a lifetime of 15 years
     )
-
 
     # Create the individual test nodes, corresponding to a system with an electricity
     # demand/sink and source
@@ -77,7 +78,7 @@ function generate_example_ss_investment(lifemode = RollingLife; discount_rate = 
             FixedProfile(0),            # Capacity in MW
             FixedProfile(10),           # Variable OPEX in EUR/MW
             FixedProfile(5),            # Fixed OPEX in EUR/MW/year
-            Dict(Power => 1),           # Output from the Node, in this case, Power
+            Dict(power => 1),           # Output from the Node, in this case, power
             [investment_data_source],   # Additional data used for adding the investment data
         ),
         RefSink(
@@ -85,16 +86,21 @@ function generate_example_ss_investment(lifemode = RollingLife; discount_rate = 
             OperationalProfile([20, 30, 40, 30]), # Demand in MW
             Dict(:surplus => FixedProfile(0), :deficit => FixedProfile(1e6)),
             # Line above: Surplus and deficit penalty for the node in EUR/MWh
-            Dict(Power => 1),           # Energy demand and corresponding ratio
+            Dict(power => 1),           # Energy demand and corresponding ratio
         ),
     ]
 
-    # Connect all nodes with the availability node for the overall energy/mass balance
+    # Connect the two nodes
+    # NOTE: This hard coding based on indexing is error prone. It is in general advised to
+    #       use a mapping dictionary to avoid any problems when introducing new technology
+    #       nodes.
     links = [
         Direct("source-demand", nodes[1], nodes[2], Linear()),
     ]
 
     # Input data structure
+    # It is also explained on
+    # https://energymodelsx.github.io/EnergyModelsBase.jl/stable/library/public/case_element/
     case = Case(T, products, [nodes, links], [[get_nodes, get_links]])
     return case, model
 end