+++ src/structures/model.jl
@@ -64,8 +64,8 @@ This abstract model type should be used when creating additional [`EnergyModel`]
 that should utilize investments.
 
 !!! note
-    Although it is declared within `EnergyModelsBase`, its concrete is only accessible if
-    `EnergyModelsInvestments` is loaded
+    Although it is declared within `EnergyModelsBase`, its concrete type is only accessible
+    if `EnergyModelsInvestments` is loaded
 
 An example for additional types is given by the inclusion of, *e.g.*, `SDDP`.
 """